refactor(embedded-mpv): drop superseded overlay hooks
