+++ libs/ui/playback/src/lib/embedded-mpv-player/index.ts
@@ -1,4 +1,3 @@
-export * from './embedded-mpv-compositor';
 export * from './embedded-mpv-overlay-visibility.service';
 export * from './embedded-mpv-player.component';
 export * from './embedded-mpv-session-controller';
