+++ tools/eslint/max-lines-baseline.mjs
@@ -120,7 +120,6 @@ export const maxLinesBaseline = [
     'libs/ui/components/src/lib/season-container/season-container.component.ts',
     'libs/ui/epg/src/lib/multi-epg/multi-epg-container.component.ts',
     'libs/ui/playback/src/lib/embedded-mpv-player/embedded-mpv-player.component.ts',
-    'libs/ui/playback/src/lib/embedded-mpv-player/embedded-mpv-session-controller.ts',
     'libs/ui/playback/src/lib/html-video-player/html-video-player.component.ts',
     'libs/ui/playback/src/lib/vjs-player/vjs-player.component.spec.ts',
     'libs/ui/playback/src/lib/vjs-player/vjs-player.component.ts',
